changed wallet id to Long from Integer

+++ src/main/java/com/kswr/wallets/api/walletsapi/controller/WalletsController.java
@@ -15,7 +15,7 @@ public class WalletsController {
     }
 
     @GetMapping("/{id}")
-    public Wallet getWalletById(@PathVariable("id") Integer id) {
+    public Wallet getWalletById(@PathVariable("id") Long id) {
         return walletsService.getWalletById(id);
     }
 }
